Splitting pins into separate files
