+++ libs/core/pxt.json
@@ -9,6 +9,9 @@
         "pxtcore.h",
         "pins.h",
         "pins.cpp",
+        "pinsAnalog.cpp",
+        "pinsDigital.cpp",
+        "pinsPWM.cpp",
         "pins.ts",
         "control.cpp",
         "i2c.cpp",
